removing depedency of core on worker
